Factor out common Statement
