+++ src/Assertion/AssertionInterface.php
@@ -6,16 +6,11 @@ namespace webignition\BasilModels\Assertion;
 
 use webignition\BasilModels\StatementInterface;
 
-interface AssertionInterface extends StatementInterface, \JsonSerializable
+interface AssertionInterface extends StatementInterface
 {
     public function getIdentifier(): string;
     public function getComparison(): string;
     public function withComparison(string $comparison): AssertionInterface;
     public function withIdentifier(string $identifier): AssertionInterface;
     public function equals(AssertionInterface $assertion): bool;
-
-    /**
-     * @return array<mixed>
-     */
-    public function jsonSerialize(): array;
 }
