+++ src/Statement.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+
+namespace webignition\BasilModels;
+
+class Statement implements StatementInterface
+{
+    private const KEY_SOURCE = 'source';
+
+    /**
+     * @var string
+     */
+    protected $source;
+
+    public function __construct(string $source)
+    {
+        $this->source = $source;
+    }
+
+    public function getSource(): string
+    {
+        return $this->source;
+    }
+
+    public function __toString(): string
+    {
+        return $this->source;
+    }
+
+    public function jsonSerialize(): array
+    {
+        return [
+            self::KEY_SOURCE => $this->source,
+        ];
+    }
+}
